Fix getopts invocation
Shell builtin `getopts` signals success whenever there is an
argument to parse (that looks like an option starting with "-"). So,
previously a command `unittest -q ...` would trigger getopts and lead
`unittest` to enter the watch mode branch.

This commit fixes that by testing for the actual option detected by
getopts.